+++ src/CssFiles/NavBarCSS.css
@@ -12,7 +12,6 @@
     margin-right: 10px; 
 }
 
-
 .navbar-nav{
     padding-right: 15px;
     padding-left: 15px;
